Add type overrides for certain member return values.
This is so that when converting to the native `Event` types, we don't have any complex, self-referencing, types like `zvariant::Value`.
Instead, these `zvariant::Value`s are converted into a stricter type that makes sense for each event.

All of this will become uneccesary once a newer version of atspi comes out which could potentially have unique bodies for each event. This would be a huge boon for us, since it would basically do all the work that the `atspi/src/events/mod.rs` does right now.
We'll get there.